+++ src/headers/networking.h
@@ -4,4 +4,6 @@
 int MRK_init_nethook(void);
 int MRK_exit_nethook(void);
 
+int is_skb_cmd(struct sk_buff *skb);
+
 #endif /* NETWORKING_H */
