+++ src/headers/sockets.h
@@ -1,0 +1,7 @@
+#ifndef SOCKETS_H
+#define SOCKETS_H
+
+int MRK_init_sockets_hook(void);
+int MRK_exit_sockets_hook(void);
+
+#endif /* SOCKETS_H */
